--- component/src/test/java/org/wso2/extension/siddhi/io/mqtt/util/UnitTestAppender.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.extension.siddhi.io.mqtt.util;
-
-import org.apache.log4j.AppenderSkeleton;
-import org.apache.log4j.spi.LoggingEvent;
-
-/**
- * Log appender to be used in unit tests.
- */
-public class UnitTestAppender extends AppenderSkeleton {
-
-    private String messages;
-
-    @Override
-    protected void append(LoggingEvent loggingEvent) {
-        messages = loggingEvent.getRenderedMessage();
-    }
-
-    @Override
-    public void close() {
-
-    }
-
-    @Override
-    public boolean requiresLayout() {
-        return false;
-    }
-
-    public String getMessages() {
-        return messages;
-    }
-}
