+++ component/src/test/resources/log4j.properties
@@ -30,7 +30,7 @@ log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%m%n
 #log4j.appender.stdout.layout.ConversionPattern=[%t] %-5p %c %x - %m%n
 
-log4j.appender.testNG=org.wso2.extension.siddhi.io.mqtt.source.LoggerAppender
+log4j.appender.testNG=io.siddhi.extension.io.mqtt.source.LoggerAppender
 log4j.appender.testNG.layout=org.apache.log4j.EnhancedPatternLayout
 log4j.appender.testNG.layout.ConversionPattern=%d{ABSOLUTE} %5p %c{1}:%L - %m%
