+++ component/src/test/resources/testng.xml
@@ -20,14 +20,14 @@
 <suite name="MqttIoTestSuit" verbose="1">
     <test name="MqttIoTest" preserve-order="true">
         <classes>
-            <class name="org.wso2.extension.siddhi.io.mqtt.sink.MqttSinkMapTest" />
-            <class name="org.wso2.extension.siddhi.io.mqtt.sink.MqttSinkQosTestCase" />
-            <class name="org.wso2.extension.siddhi.io.mqtt.sink.MqttSinkTestCase" />
-            <class name="org.wso2.extension.siddhi.io.mqtt.sink.MqttPublishTestCase" />
+            <class name="io.siddhi.extension.io.mqtt.sink.MqttSinkMapTest" />
+            <class name="io.siddhi.extension.io.mqtt.sink.MqttSinkQosTestCase" />
+            <class name="io.siddhi.extension.io.mqtt.sink.MqttSinkTestCase" />
+            <class name="io.siddhi.extension.io.mqtt.sink.MqttPublishTestCase" />
 
-            <class name="org.wso2.extension.siddhi.io.mqtt.source.MqttSourceQosTest" />
-            <class name="org.wso2.extension.siddhi.io.mqtt.source.MqttSourceMapTest" />
-            <class name="org.wso2.extension.siddhi.io.mqtt.source.MqttSourceTestCase" />
+            <class name="io.siddhi.extension.io.mqtt.source.MqttSourceQosTest" />
+            <class name="io.siddhi.extension.io.mqtt.source.MqttSourceMapTest" />
+            <class name="io.siddhi.extension.io.mqtt.source.MqttSourceTestCase" />
 
         </classes>
     </test>
